+++ engine/src/geometry/g_common.c
@@ -9,7 +9,7 @@
 
 // Returns 0 on success, 1 on failure
 int upload_vertices(
-    SDL_GPUDevice* device, const void* vertices, size_t vertices_size,
+    SDL_GPUDevice* device, const void* vertices, Uint64 vertices_size,
     SDL_GPUBuffer** vbo_out
 ) {
     SDL_GPUBufferCreateInfo vbo_info = {
@@ -81,7 +81,7 @@ int upload_vertices(
 
 // Returns 0 on success, 1 on failure
 int upload_indices(
-    SDL_GPUDevice* device, const void* indices, size_t indices_size,
+    SDL_GPUDevice* device, const void* indices, Uint64 indices_size,
     SDL_GPUBuffer** ibo_out
 ) {
     SDL_GPUBufferCreateInfo ibo_info = {
@@ -152,7 +152,7 @@ int upload_indices(
 }
 
 void compute_vertex_normals(
-    float* vertices, int num_vertices, const uint16_t* indices,
+    float* vertices, int num_vertices, const Uint16* indices,
     int num_indices, int stride, int pos_offset, int norm_offset
 ) {
     // Accumulate normals per vertex
